adding planning api

+++ costos/statistics.py
@@ -415,13 +415,13 @@ def get_planning(request):
             "product_name": "TOTALES",
             "year": "2024",
             "month": "Enero",
-            "total": sum([d.get("total", 0) for d in planning]),
+            "total": float(sum([d.get("total", 0) for d in planning])),
             "precio": 0,
-            "total_venta": sum([d.get("total_venta", 0) if d.get("total_venta") != '' and d.get("total_venta") is not None else 0 for d in planning]),
-            "total_actual": sum([d.get("total_actual", 0) if d.get("total_actual") != '' and d.get("total_actual") is not None else 0 for d in planning]),
-            "total_venta_actual": sum([d.get("total_venta_actual", 0) if d.get("total_venta_actual") != '' and d.get("total_venta_actual") is not None else 0 for d in planning]),
-            "costo_producto": sum([d.get("costo_producto", 0) if d.get("costo_producto") != '' and d.get("costo_producto") is not None else 0 for d in planning]),
-            "ganancia": sum([d.get("ganancia") if d.get("ganancia") != '' and d.get("ganancia") is not None else 0 for d in planning]) / sum([1 if d.get("ganancia") != '' and d.get("ganancia") is not None else 0 for d in planning]),
+            "total_venta": float(sum([d.get("total_venta", 0) if d.get("total_venta") != '' and d.get("total_venta") is not None else 0 for d in planning])),
+            "total_actual": float(sum([d.get("total_actual", 0) if d.get("total_actual") != '' and d.get("total_actual") is not None else 0 for d in planning])),
+            "total_venta_actual": float(sum([d.get("total_venta_actual", 0) if d.get("total_venta_actual") != '' and d.get("total_venta_actual") is not None else 0 for d in planning])),
+            "costo_producto": float(sum([d.get("costo_producto", 0) if d.get("costo_producto") != '' and d.get("costo_producto") is not None else 0 for d in planning])),
+            "ganancia": float(sum([d.get("ganancia") if d.get("ganancia") != '' and d.get("ganancia") is not None else 0 for d in planning]) / sum([1 if d.get("ganancia") != '' and d.get("ganancia") is not None else 0 for d in planning])),
         })
 
     return JsonResponse(planning, safe=False)
